// packages/nars-client/src/Schema.d.ts
export namespace google_mirror {

    namespace protobuf {

        interface IStruct {
            fields?: ({ [k: string]: google_mirror.protobuf.IValue }|null);
        }

        class Struct implements IStruct {
            constructor(properties?: google_mirror.protobuf.IStruct);
            public fields: { [k: string]: google_mirror.protobuf.IValue };
            public static create(properties?: google_mirror.protobuf.IStruct): google_mirror.protobuf.Struct;
            public static encode(message: google_mirror.protobuf.IStruct, writer?: $protobuf.Writer): $protobuf.Writer;
            public static encodeDelimited(message: google_mirror.protobuf.IStruct, writer?: $protobuf.Writer): $protobuf.Writer;
            public static decode(reader: ($protobuf.Reader|Uint8Array), length?: number): google_mirror.protobuf.Struct;
            public static decodeDelimited(reader: ($protobuf.Reader|Uint8Array)): google_mirror.protobuf.Struct;
            public static verify(message: { [k: string]: any }): (string|null);
            public static fromObject(object: { [k: string]: any }): google_mirror.protobuf.Struct;
            public static toObject(message: google_mirror.protobuf.Struct, options?: $protobuf.IConversionOptions): { [k: string]: any };
            public toJSON(): { [k: string]: any };
        }

        interface IValue {
            nullValue?: (google_mirror.protobuf.NullValue|null);
            numberValue?: (number|null);
            stringValue?: (string|null);
            boolValue?: (boolean|null);
            structValue?: (google_mirror.protobuf.IStruct|null);
            listValue?: (google_mirror.protobuf.IListValue|null);
        }

        class Value implements IValue {
            constructor(properties?: google_mirror.protobuf.IValue);
            public nullValue: google_mirror.protobuf.NullValue;
            public numberValue: number;
            public stringValue: string;
            public boolValue: boolean;
            public structValue?: (google_mirror.protobuf.IStruct|null);
            public listValue?: (google_mirror.protobuf.IListValue|null);
            public kind?: ("nullValue"|"numberValue"|"stringValue"|"boolValue"|"structValue"|"listValue");
            public static create(properties?: google_mirror.protobuf.IValue): google_mirror.protobuf.Value;
            public static encode(message: google_mirror.protobuf.IValue, writer?: $protobuf.Writer): $protobuf.Writer;
            public static encodeDelimited(message: google_mirror.protobuf.IValue, writer?: $protobuf.Writer): $protobuf.Writer;
            public static decode(reader: ($protobuf.Reader|Uint8Array), length?: number): google_mirror.protobuf.Value;
            public static decodeDelimited(reader: ($protobuf.Reader|Uint8Array)): google_mirror.protobuf.Value;
            public static verify(message: { [k: string]: any }): (string|null);
            public static fromObject(object: { [k: string]: any }): google_mirror.protobuf.Value;
            public static toObject(message: google_mirror.protobuf.Value, options?: $protobuf.IConversionOptions): { [k: string]: any };
            public toJSON(): { [k: string]: any };
        }

        enum NullValue {
            NULL_VALUE = 0
        }

        interface IListValue {
            values?: (google_mirror.protobuf.IValue[]|null);
        }

        class ListValue implements IListValue {
            constructor(properties?: google_mirror.protobuf.IListValue);
            public values: google_mirror.protobuf.IValue[];
            public static create(properties?: google_mirror.protobuf.IListValue): google_mirror.protobuf.ListValue;
            public static encode(message: google_mirror.protobuf.IListValue, writer?: $protobuf.Writer): $protobuf.Writer;
            public static encodeDelimited(message: google_mirror.protobuf.IListValue, writer?: $protobuf.Writer): $protobuf.Writer;
            public static decode(reader: ($protobuf.Reader|Uint8Array), length?: number): google_mirror.protobuf.ListValue;
            public static decodeDelimited(reader: ($protobuf.Reader|Uint8Array)): google_mirror.protobuf.ListValue;
            public static verify(message: { [k: string]: any }): (string|null);
            public static fromObject(object: { [k: string]: any }): google_mirror.protobuf.ListValue;
            public static toObject(message: google_mirror.protobuf.ListValue, options?: $protobuf.IConversionOptions): { [k: string]: any };
            public toJSON(): { [k: string]: any };
        }
    }
}